backend_new/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# Spotify API
spotipy

# Environment Variables
python-dotenv

# Data Validation
pydantic

# Sentiment Analysis
vaderSentiment

# AI/LLM Services
groq

# Audio Transcription
deepgram-sdk

# HTTP Requests (for future use)
requests